Update dev.yaml
I will create new branches for each post for the foreseeable future (deleting them as they are merged into main). I want to make sure that each branch individually functions according ot the server requirements. As a result, any new branch for a new post needs to begin with `posts_` in order to make sure that this github action will run on them.